:bug: Fix: 코드 문법 수정

+++ src/main/kotlin/team/sparta/onehouronemeal/domain/recipe/service/v1/RecipeService.kt
@@ -1,5 +1,7 @@
 package team.sparta.onehouronemeal.domain.recipe.service.v1
 
+import jakarta.transaction.Transactional
+import org.springframework.data.repository.findByIdOrNull
 import org.springframework.stereotype.Service
 import team.sparta.onehouronemeal.domain.course.repository.v1.CourseRepository
 import team.sparta.onehouronemeal.domain.recipe.dto.v1.CreateRecipeRequest
@@ -7,6 +9,7 @@ import team.sparta.onehouronemeal.domain.recipe.dto.v1.RecipeResponse
 import team.sparta.onehouronemeal.domain.recipe.dto.v1.UpdateRecipeRequest
 import team.sparta.onehouronemeal.domain.recipe.model.v1.Recipe
 import team.sparta.onehouronemeal.domain.recipe.repository.v1.RecipeRepository
+import team.sparta.onehouronemeal.exception.ModelNotFoundException
 
 @Service
 class RecipeService(
@@ -14,21 +17,19 @@ class RecipeService(
     private val courseRepository: CourseRepository,
 ) {
     fun getAllRecipeList(courseId: Long): List<RecipeResponse> {
-        val course = courseRepository.findById(courseId)
-            .orElseThrow { RuntimeException("Course not found") }
-        return recipeRepository.findByCourse(course).map { RecipeResponse.from(it) }
+        if (!courseRepository.existsById(courseId)) throw ModelNotFoundException("Course", courseId)
+        return recipeRepository.findAllByCourseId(courseId).map { RecipeResponse.from(it) }
     }
 
     fun getRecipeById(courseId: Long, recipeId: Long): RecipeResponse {
-        val course = courseRepository.findById(courseId).orElseThrow { RuntimeException("Course not found") }
+        if (!courseRepository.existsById(courseId)) throw ModelNotFoundException("Course", courseId)
         val recipe =
-            recipeRepository.findByCourseAndId(course, recipeId).orElseThrow { RuntimeException("Recipe not found") }
+            recipeRepository.findByCourseIdAndId(courseId, recipeId) ?: throw ModelNotFoundException("Recipe", recipeId)
         return RecipeResponse.from(recipe)
     }
 
     fun createRecipe(courseId: Long, request: CreateRecipeRequest): RecipeResponse {
-        val course = courseRepository.findById(courseId)
-            .orElseThrow { RuntimeException("Course not found") }
+        val course = courseRepository.findByIdOrNull(courseId) ?: throw ModelNotFoundException("Course", courseId)
         val recipe = Recipe(
             course = course,
             title = request.title,
@@ -38,24 +39,24 @@ class RecipeService(
         return RecipeResponse.from(recipeRepository.save(recipe))
     }
 
+    @Transactional
     fun updateRecipe(courseId: Long, recipeId: Long, request: UpdateRecipeRequest): RecipeResponse {
-        val course = courseRepository.findById(courseId)
-            .orElseThrow { RuntimeException("Course not found") }
-        val recipe = recipeRepository.findByCourseAndId(course, recipeId)
-            .orElseThrow { RuntimeException("Recipe not found") }
+        if (!courseRepository.existsById(courseId)) throw ModelNotFoundException("Course", courseId)
+        val recipe =
+            recipeRepository.findByCourseIdAndId(courseId, recipeId) ?: throw ModelNotFoundException("Recipe", recipeId)
 
         recipe.title = request.title
         recipe.describe = request.describe
         recipe.videoUrl = request.videoUrl
 
-        return RecipeResponse.from(recipeRepository.save(recipe))
+        return RecipeResponse.from(recipe)
     }
 
+    @Transactional
     fun deleteRecipe(courseId: Long, recipeId: Long) {
-        val course = courseRepository.findById(courseId)
-            .orElseThrow { RuntimeException("Course not found") }
-        val recipe = recipeRepository.findByCourseAndId(course, recipeId)
-            .orElseThrow { RuntimeException("Recipe not found") }
+        if (!courseRepository.existsById(courseId)) throw ModelNotFoundException("Course", courseId)
+        val recipe =
+            recipeRepository.findByCourseIdAndId(courseId, recipeId) ?: throw ModelNotFoundException("Recipe", recipeId)
 
         recipeRepository.delete(recipe)
     }
